telegram-stock-bot/requirements.txt
telethon>=1.36.0
openai>=1.12.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
